# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CodeMind API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/auth/recovery-key/generate": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "API endpoint for /api/auth/recovery-key/generate",
        "description": "Tự động sinh tài liệu cho POST /api/auth/recovery-key/generate. Hãy cập nhật mô tả chi tiết sau.",
        "responses": {
          "200": {
            "description": "Thành công"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        }
      }
    },
...
    "/api/blog/posts/[slug]/like": {
      "post": {
        "tags": [
          "Blog"
        ],
        "summary": "API endpoint for /api/blog/posts/[slug]/like",
        "description": "Tự động sinh tài liệu cho POST /api/blog/posts/[slug]/like. Hãy cập nhật mô tả chi tiết sau.",
        "responses": {
          "200": {
            "description": "Thành công"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        }
      }
    },
...
    "/api/blog/posts/[slug]/share": {
      "post": {
        "tags": [
          "Blog"
        ],
        "summary": "API endpoint for /api/blog/posts/[slug]/share",
        "description": "Tự động sinh tài liệu cho POST /api/blog/posts/[slug]/share. Hãy cập nhật mô tả chi tiết sau.",
        "responses": {
          "200": {
            "description": "Thành công"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        }
      }
    },
...
    "/api/cv/parse": {
      "post": {
        "tags": [
          "Cv"
        ],
        "summary": "API endpoint for /api/cv/parse",
        "description": "Tự động sinh tài liệu cho POST /api/cv/parse. Hãy cập nhật mô tả chi tiết sau.",
        "responses": {
          "200": {
            "description": "Thành công"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        }
      }
    },
...
    "/api/lessons/[lessonId]/questions": {
      "get": {
        "tags": [
          "Lessons"
        ],
        "summary": "API endpoint for /api/lessons/[lessonId]/questions",
        "description": "Tự động sinh tài liệu cho GET /api/lessons/[lessonId]/questions. Hãy cập nhật mô tả chi tiết sau.",
        "responses": {
          "200": {
            "description": "Thành công"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        }
      },
      "post": {
        "tags": [
          "Lessons"
        ],
        "summary": "API endpoint for /api/lessons/[lessonId]/questions",
        "description": "Tự động sinh tài liệu cho POST /api/lessons/[lessonId]/questions. Hãy cập nhật mô tả chi tiết sau.",
        "responses": {
          "200": {
            "description": "Thành công"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        }
      }
    },
...
  },
...
}